[SKP-182] refac : pushUtils 분리
